requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.3
IMDbPY==2022.7.9
pandas==1.4.2
requests==2.27.1
SQLAlchemy==1.4.39
